Refactor, correct timing correction (for most sensors)
