+++ nebula_decoders/include/nebula_decoders/nebula_decoders_hesai/decoders/hesai_packet.hpp
@@ -14,6 +14,8 @@ namespace drivers
 namespace hesai_packet
 {
 
+// FIXME(mojomex) This is a workaround for the compiler being pedantic about casting `enum class`s
+// to their underlying type
 namespace return_mode
 {
 enum ReturnMode {
@@ -170,105 +172,6 @@ struct Body
   BlockT blocks[BlockN];
 };
 
-struct Tail128E3X
-{
-  uint8_t reserved1[9];
-  uint16_t azimuth_state;
-  uint8_t operational_state;
-  uint8_t return_mode;
-  uint16_t motor_speed;
-  DateTime<1900> date_time;
-  uint32_t timestamp;
-  uint8_t factory_information;
-
-  /* Ignored optional fields */
-
-  //uint32_t udp_sequence;
-
-  //uint16_t imu_temperature;
-  //uint16_t imu_acceleration_unit;
-  //uint16_t imu_angular_velocity_unit;
-  //uint32_t imu_timestamp;
-  //uint16_t imu_x_axis_acceleration;
-  //uint16_t imu_y_axis_acceleration;
-  //uint16_t imu_z_axis_acceleration;
-  //uint16_t imu_x_axis_angular_velocity;
-  //uint16_t imu_y_axis_angular_velocity;
-  //uint16_t imu_z_axis_angular_velocity;
-
-  //uint32_t crc_tail;
-};
-
-struct TailXT32
-{
-  uint8_t reserved1[10];
-  uint8_t return_mode;
-  uint16_t motor_speed;
-  DateTime<1900> date_time;
-  uint32_t timestamp;
-  uint8_t factory_information;
-};
-
-struct Tail40P
-{
-  uint8_t reserved1[5];
-  uint8_t high_temperature_shutdown_flag;
-  uint8_t reserved2[2];
-  uint16_t motor_speed;
-  uint32_t timestamp;
-  uint8_t return_mode;
-  uint8_t factory_information;
-  DateTime<2000> date_time;
-
-  /* Ignored optional fields */
-
-  //uint32_t udp_sequence;
-};
-
-struct TailQT128C2X
-{
-  uint8_t reserved1[5];
-  uint8_t mode_flag;
-  uint8_t reserved2[6];
-  uint8_t return_mode;
-  uint16_t motor_speed;
-  DateTime<1900> date_time;
-  uint32_t timestamp;
-  uint8_t factory_information;
-  
-  /* Ignored optional fields */
-
-  //uint32_t udp_sequence;
-  //uint32_t crc_tail;
-};
-
-struct TailAT128E2X
-{
-  uint8_t reserved1[6];
-  uint8_t high_temperature_shutdown_flag;
-  uint8_t reserved2[11];
-  uint16_t motor_speed;
-  uint32_t timestamp;
-  uint8_t return_mode;
-  uint8_t factory_information;
-  SecondsSinceEpoch date_time;
-  
-  /* Ignored optional fields */
-
-  //uint32_t udp_sequence;
-  //uint32_t crc_tail;
-};
-
-struct TailQT64
-{
-  uint8_t reserved1[10];
-  uint16_t motor_speed;
-  uint32_t timestamp;
-  uint8_t return_mode;
-  uint8_t factory_information;
-  DateTime<1900> date_time;
-};
-
 /// @brief Base struct for all Hesai packets. This struct is not allowed to have any non-static
 /// members, otherwise memory layout is not guaranteed for the derived structs.
 /// @tparam nBlocks The number of blocks in the packet
@@ -283,106 +186,6 @@ struct PacketBase
   static constexpr size_t N_CHANNELS = nChannels;
   static constexpr size_t MAX_RETURNS = maxReturns;
   static constexpr size_t DEGREE_SUBDIVISIONS = degreeSubdivisions;
-};
-
-struct Packet128E3X : public PacketBase<2, 128, 2, 100>
-{
-  typedef Body<Block<Unit3B, Packet128E3X::N_CHANNELS>, Packet128E3X::N_BLOCKS> body_t;
-  typedef AngleCorrectorCalibrationBased<
-    Packet128E3X::N_CHANNELS, Packet128E3X::DEGREE_SUBDIVISIONS>
-    angle_decoder_t;
-  Header12B header;
-  body_t body;
-  uint32_t crc_body;
-  FunctionalSafety fs;
-  Tail128E3X tail;
-
-  // FIXME(mojomex) uint8_t cyber_security[32]; is added optionally
-};
-
-struct PacketXT32 : public PacketBase<8, 32, 2, 100>
-{
-  typedef Body<Block<Unit4B, PacketXT32::N_CHANNELS>, PacketXT32::N_BLOCKS> body_t;
-  typedef AngleCorrectorCalibrationBased<PacketXT32::N_CHANNELS, PacketXT32::DEGREE_SUBDIVISIONS>
-    angle_decoder_t;
-  Header12B header;
-  body_t body;
-  TailXT32 tail;
-  uint32_t udp_sequence;
-};
-
-typedef TailXT32 TailXT32M2X;
-struct PacketXT32M2X : public PacketBase<6, 32, 3, 100>
-{
-  typedef Body<Block<Unit4B, PacketXT32M2X::N_CHANNELS>, PacketXT32M2X::N_BLOCKS> body_t;
-  typedef AngleCorrectorCalibrationBased<
-    PacketXT32M2X::N_CHANNELS, PacketXT32M2X::DEGREE_SUBDIVISIONS>
-    angle_decoder_t;
-  Header12B header;
-  body_t body;
-  TailXT32M2X tail;
-  uint32_t udp_sequence;
-};
-
-struct PacketQT128C2X : public PacketBase<2, 128, 2, 100>
-{
-  typedef Body<Block<Unit4B, PacketQT128C2X::N_CHANNELS>, PacketQT128C2X::N_BLOCKS> body_t;
-  typedef AngleCorrectorCalibrationBased<
-    PacketQT128C2X::N_CHANNELS, PacketQT128C2X::DEGREE_SUBDIVISIONS>
-    angle_decoder_t;
-  Header12B header;
-  body_t body;
-  uint32_t crc_body;
-  FunctionalSafety fs;
-  TailQT128C2X tail;
-  /// @brief Always present, all zeros by default
-  uint8_t cyber_security[32];
-};
-
-struct PacketAT128E2X : public PacketBase<2, 128, 2, 100 * 256>
-{
-  typedef Body<FineAzimuthBlock<Unit4B, PacketAT128E2X::N_CHANNELS>, PacketAT128E2X::N_BLOCKS>
-    body_t;
-  typedef AngleCorrectorCorrectionBased<
-    PacketAT128E2X::N_CHANNELS, PacketAT128E2X::DEGREE_SUBDIVISIONS>
-    angle_decoder_t;
-  Header12B header;
-  body_t body;
-  uint32_t crc_body;
-  TailAT128E2X tail;
-  /// @brief Always present, not supported according to datasheet
-  uint8_t cyber_security[32];
-};
-
-struct PacketQT64 : public PacketBase<4, 64, 2, 100>
-{
-  typedef Body<Block<Unit4B, PacketQT64::N_CHANNELS>, PacketQT64::N_BLOCKS> body_t;
-  typedef AngleCorrectorCalibrationBased<PacketQT64::N_CHANNELS, PacketQT64::DEGREE_SUBDIVISIONS>
-    angle_decoder_t;
-  Header12B header;
-  body_t body;
-  TailQT64 tail;
-  uint32_t udp_sequence;
-};
-
-typedef Tail40P Tail64;
-struct Packet64 : public PacketBase<6, 64, 2, 100>
-{
-  typedef Body<Block<Unit3B, Packet64::N_CHANNELS>, Packet64::N_BLOCKS> body_t;
-  typedef AngleCorrectorCalibrationBased<Packet64::N_CHANNELS, Packet64::DEGREE_SUBDIVISIONS>
-    angle_decoder_t;
-  Header8B header;
-  body_t body;
-  Tail64 tail;
-};
-
-struct Packet40P : public PacketBase<10, 40, 2, 100>
-{
-  typedef Body<SOBBlock<Unit3B, Packet40P::N_CHANNELS>, Packet40P::N_BLOCKS> body_t;
-  typedef AngleCorrectorCalibrationBased<Packet40P::N_CHANNELS, Packet40P::DEGREE_SUBDIVISIONS>
-    angle_decoder_t;
-  body_t body;
-  Tail40P tail;
 };
 
 #pragma pack(pop)
@@ -401,6 +204,39 @@ uint64_t get_timestamp_ns(const PacketT & packet);
 
 template <typename PacketT>
 uint8_t get_dis_unit(const PacketT & packet);
+
+int get_n_returns(uint8_t return_mode)
+{
+  switch (return_mode) {
+    case return_mode::SINGLE_FIRST:
+    case return_mode::SINGLE_SECOND:
+    case return_mode::SINGLE_STRONGEST:
+    case return_mode::SINGLE_LAST:
+      return 1;
+    case return_mode::DUAL_LAST_STRONGEST:
+    case return_mode::DUAL_FIRST_SECOND:
+    case return_mode::DUAL_FIRST_LAST:
+    case return_mode::DUAL_FIRST_STRONGEST:
+    case return_mode::DUAL_STRONGEST_SECONDSTRONGEST:
+      return 2;
+    case return_mode::TRIPLE_FIRST_LAST_STRONGEST:
+      return 3;
+    default:
+      throw std::runtime_error("Unknown return mode");
+  }
+}
+
+template <typename PacketT>
+uint64_t get_timestamp_ns(const PacketT & packet)
+{
+  return packet.tail.date_time.get_seconds() * 1000000000 + packet.tail.timestamp * 1000;
+}
+
+template <typename PacketT>
+uint8_t get_dis_unit(const PacketT & packet)
+{
+  return packet.header.dis_unit;
+}
 
 }  // namespace hesai_packet
 }  // namespace drivers
